[bugfix] out/wl: draw bottom instead of wallpaper
Since swaybg, which I based the source code on occupies the "background"
layer, I cannot use it to draw XAVA, because that will cause the window
to spawn behing swaybg, which ruins the entire experience for everyone.

So instead there is, thankfully, a "bottom" layer in the protocol that
does the job just right. Thanks sway devs :)