[design fix] no externs beyond this point!
When XAVA switched to the dynamic module style of handling features, the
dynamic loader has a extremely hard time processing externs and such.
A typical example would be on Windows where symbol sharing between
dynamically loaded variables and the program executable is outright
outlawed - it'll crash if you try to access internal variables using a
library.

The only way this could be fixed is by switching to the new design which
only allows static and function-passed variables preventing issues such
as above.