[QOL] updated config and defaults for Wayland
XAVA's XDG implementation is lacking in both features and stability
which is why it has been disabled as the default. As for ZWLR, it has
enough functionality to "work" at least, it's getting promoted despite
the incompatibilty with Mutter or other compositors. In reality those
users should use x11 instead, since their XWayland implementation
actually works as opposed to wlroots.