[fix] Made the logging system a shared library
It's because loading symbols off of the binary executing the library is
difficult to say the least (at least on PE+ executables used on Windows).
This measure is a design workaround that avoids this problem by making
the shared functions to be exposed by library rather in the main
executable.

Does this increase load times... yes. But it's less hell than statically
linking together code that might not be compatible, so weight your
options, I guess.

Anyways, Tschuss!