[feature] Wayland EGL support
Following my "modules" design, it is activated by switching the output
mode to "wayland_egl". Once that's done a seperate code path is run
(determined at compile time).

Anyway, for now, this mode is pretty basic, only drawing a red
background, but hey, it's important that we got it off the ground.

Now the only thing that's left to do is to:
 1) implement shaders
 2) add default shaders
 3) implement VBOs or any other type of vertex data transfer
 4) make it work nicely

However, I've noticed that glClear is stupid expensive on a GPU. Might
have to look into that.

Another idea is to switch everything to EGL as it's only a sub-set of
the real OpenGL. And we really don't need the full spec (at least for
now).