[thing] moved all EGL code to a shared place
Since this is the renderer I'm most likely sticking with, I am moving
all of the "just raw EGL" code to a seperate place so I don't have to
rewrite it. And as you may see, there's a crap load of code to deal
with. However, this code is thankfully all universal and only like 20%
of the entire EGL code base is platform specific (at least for Wayland)

My hope is that I can reuse this while porting the EGL renderer to other
platforms, such as X11, Win32 or even Cocoa (maybe if Apple wasn't
selling overpriced stuff first ;) )