[bugfix] fix glClear color blending
When using X11 or Wayland with EGL, the backgrond color blending is
broken in such a fashion where the colors are added anyway if their
opacity is zero to the current color, which causes an undesired blended
image result. In order to remedy this, you can multiply the individual
color factors by the opacity. That way the colors that are
added/subtracted are way more subtle than without this blending mode and
sucessfully creates our desired result.

I hate how the color blending works on *NIX systems, but it's something
you'll have to live with, I guess.