[removal] remove OpenGL functions from graphical.h
Graphical.h is supposed to be platform-agnostic, so that's why I'm
ridding it of any OpenGL code. Only code that can be generic and useful
in any situation should be put in graphical.h (because that's it's whole
purpose).

For OpenGL (or any API) specific solution a shared source file found in
the shared folder should be used!

This is part 1 towards migrating the entirety of WGL to a shared
reusable OGL codebase.