[feature] Merge all the OpenGL codebases (#14)
Since GLES is just a subset of full desktop OpenGL, I thought it might
be possible to merge and run the same code instead of forcing GLES to
run a platform that it wasn't designed for.

This prompted a new gl_shared.c and gl_shared.h file to be made. These
files contain what is the shared code between the GLES and OpenGL
outputs. This allows for any changes to apply across both platforms (for
convenience and ease of use sakes). That means that any changes for the
system should work across both systems. However there's a slight
limitation: Since both codebases are the same (including the shaders)
the developer must make sure that his/her shaders work on both GLES and
full desktop OpenGL.

Notes:
	Shaders are written in accordance to the OpenGL ES 2.0 standard
and are now placed in the "gl" instead of the "egl" directory.
	Any platform specific stuff goes into the output method's code,
any API specific stuff goes into the API's source code (eg. OpenGL ==
gl.c, etc.), that includes stuff like GLEW which are necessary to
prevent hair pulling when you try to use any modern OpenGL functions.
	Probably forgot to mention something else important, but who
cares I want to push this thing to "master" as fast as I can.

PS: IT FINALLY WORKS AAAAAAAAAAAAA