[bugfix] fix memory corruption in log function
As previous nik predicted, that design led to memory corruption (and
specifically on Windows, where it's the most difficult to find, thought
SIGTRAP helped a FUCKTON here).

Imposing a charater limit instead and using a stack allocated design
helps with the memory corruption issues. Yes, it might suck; but who the
fuck needs a 500+ char function name.