[change] legacying a fuckton of code
Basically I want this project to look the same no matter what setup you
use, that means using a singular codebase aka. using the same GL/EGL
morphed renderer.

The rest of the code will be moved to a "unsupported" branch and will
contain essentially just software renderers that have no GPU support.
The name "unsupported" is to discern between "supported" and
"unsupported" features. By saying an output module is "unsupported" that
means that the module would not support the same features as the
"supported" one. They are left in because they provide a good enough
visualization platform for those without GPU acceleration (such as ARM
dev boards and other embedded devices).