[feature] allow the visalizer to hold it's size
When we resize the vizualizer beyond its defined width and height, it'll
keep expanding definitely. But what if we instead wanted the visualizer
to stay within the window, confined to the size that it was given.

Well this feature adds that, and it's enabled by enabling:
hold_size = true in the [window] section.

Enjoy :)

PS: this is the kind of freedom that the last commit allows me to do as
a developer