[change] work on moving everything to a god handle
This is just a start to moving every piece of code to a "god" handle.
In this "god" handle, in theory, I'd be able to manage resources way
easier because I shouldn't have to rely on sharing data with all of the
functions, instead all of them follow a "class-like" structure.

Yes, I'm doing OOP stuff in C, deal with it.

The first part was to move the module code to be inherited from the
"god" handle.