fuck it, we ball
this commit will almost certainly break shit and i cannot be bothered to
keep this work unaccounted for, so I'd rather just push and have the
issues roll in instead.

Yes, this is a minor refactor of the IO/shared function code for
utility functions used by XAVA's modules.

Trying to squash #67 with this one